Changed producer interface to work with long instead of int

+++ rxjava-core/src/main/java/rx/Producer.java
@@ -17,6 +17,6 @@ package rx;
 
 public interface Producer {
 
-    public void request(int n);
+    public void request(long n);
 
 }
